requirements.txt
absl-py==1.4.0
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854164373/work
cachetools==5.3.1
certifi==2022.12.7
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1636046055389/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1678108872112/work
contourpy==1.0.7
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1652967134882/work
cycler==0.11.0
Distance==0.1.3
einops @ file:///home/conda/feedstock_root/build_artifacts/einops_1681884646122/work
filelock==3.12.0
fonttools==4.39.3
fsspec==2023.5.0
google-auth==2.19.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
huggingface-hub==0.14.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
importlib-metadata==6.6.0
importlib-resources==5.12.0
joblib==1.2.0
jsonlines @ file:///tmp/build/80754af9/jsonlines_1609785323125/work
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.4.4
lmdb==1.4.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
networkx==3.1
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
oauthlib==3.2.2
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.7.0.72
packaging==23.1
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1621287979692/work
prettytable==3.7.0
protobuf==4.23.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1663846997386/work
pyparsing==3.0.9
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.5.5
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1680286922386/work
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tensorboard==2.13.0
tensorboard-data-server==0.7.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.12.1
torchnet==0.0.4
torchtoolbox==0.1.8.2
torchvision==0.13.1
tornado==6.3.1
tqdm==4.65.0
transformers==4.29.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1678635778344/work
visdom==0.2.4
vit-pytorch @ file:///home/conda/feedstock_root/build_artifacts/vit-pytorch_1640225907683/work
wcwidth==0.2.6
websocket-client==1.5.1
Werkzeug==2.3.4
zipp==3.15.0
